:memo: README: Add initial structure from existing documentation
This commit updates the Kotani Pay API repository README in compliance
to the UNICEF Venture Fund workplan requirements for open source. All
content in this README is sourced from other project docs and websites.
Links are added to add better visibility to the different pieces. The
structure is provided to give an idea of what kind of content generally
is included.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>